+++ src/main.cpp
@@ -39,6 +39,7 @@ with this program; if not, write to the Free Software Foundation, Inc.,
 #include "config.h"
 #include "player.h"
 #include "porting.h"
+#include "network/socket.h"
 #if USE_CURSES
 	#include "terminal_chat_console.h"
 #endif
